backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
pymongo==4.5.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
Pillow==10.0.0
numpy==1.24.3
tensorflow==2.13.0
opencv-python==4.8.0.74
scikit-learn==1.3.0
Werkzeug==2.3.6
python-dateutil==2.8.2